Break MapBox into smaller pieces
